Added methods to controller and BprocNode to check nodes to make certain they aren't hung....if hung then skips
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/DJob/trunk@3761 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ DistribJob/lib/perl/BprocNode.pm
@@ -10,13 +10,15 @@ use strict;
 sub new {
     my ($class, $nodeNum, $nodeDir, $slotcount) = @_;
     my $self = &DJob::DistribJob::Node::new($class, $nodeNum, $nodeDir, $slotcount);
+    return undef unless $self->_initNodeDir();  
     $self->{nodeDir} = $nodeDir;
-    $self->_initNodeDir();  
     return $self;
 }
 
 sub _initNodeDir {
     my ($self) = @_;
+
+    return 0 unless ($self->checkNode());
 
     if ($self->_fileExists($self->{nodeDir})) {
 	$self->runCmd("/bin/rm -r $self->{nodeDir}");
@@ -28,6 +30,7 @@ sub _initNodeDir {
 	    if ($try++ > 3);
 	$self->runCmd("mkdir -p $self->{nodeDir}");
     }
+    return 1;
 }
 
 sub runCmd {
@@ -53,6 +56,54 @@ sub _fileExists {
     my $test = `bpsh -n $self->{nodeNum} find $file 2> /dev/null`;
     return 1 if $test =~ /$file/;
   }
+}
+
+sub checkNode {
+  my($self) = @_;
+  my $node = $self->getNum();
+  ##First run diagnose to see if load is too high
+  my @diag = `diagnose -n node$node`;
+  foreach my $line (@diag){
+    if($line =~ /node$node.*DEF\s*(\d+)/){
+#      print STDERR $line;
+#      print STDERR "  processes: $1\n";
+      return 0 if $1 > 3;
+    }
+  }
+  ##need to try these things and time out!!
+#  return 0 unless $self->tryCommand("echo cmd","cmd",1);
+  return 0 unless $self->tryCommand("bpsh -n $node hostname","node$node",1);
+  return $self->tryCommand("bpsh -n $node ls $ENV{HOME}","No such",0);
+}
+
+sub tryCommand {
+  my($self,$cmd,$retVal,$test) = @_;
+  my $pid;
+  my $hit = 0;
+  eval {
+    local $SIG{'ALRM'} = sub { die "alarm\n"; };
+    alarm(5);
+    $pid = open(BPSH, "$cmd |");
+    if (defined($pid)) {
+      while (<BPSH>) {
+        if(/$retVal/){
+          $hit = 1;
+        }
+      }
+      alarm(0);
+    } else {
+      die "bpsh error";
+    }
+  };
+  if($@){
+    if ($@ =~ /alarm/) {
+      kill('KILL', $pid);
+#      print STDERR "KILL: $@\n";
+    }else{
+      print STDERR "Something bad happened $@\n";
+    }
+  }
+  return $hit ? $test : !$test; 
 }
 
 1;
